Profile (#37)
Switch from sys.settrace to sys.setprofile

The profile part has less overhead, as it only gets called for call and return.
Moreover, we do not need to care about threads, as Score-P does this for us.
In theory, it is also possible to catch calls to builtin C functions. However, this is not implemented.
Also added a minimal async example.