Adapt to numpy NEP-18 (#65)
Together with the introduction of [NEP-18](https://numpy.org/neps/nep-0018-array-function-protocol.html) some numpy functions appeared the name `None:dot`. `None` is misleading. The solution discussed in #63 is to make the reference to numpy and the related function `__array_function__` explicit.

As we are already started using co_filename, we changed to co functions as well.